ahdit/requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Environment Management
python-dotenv==1.0.0

# Additional utilities
click==8.1.7